"Connect mailbox" onboarding for delegated OAuth2 flows instead of manual refresh-token paste: rename view and menu 5546
