+++ jmix-email/email-flowui/src/main/resources/io/jmix/emailflowui/view/emailconnection/email-connection-view.xml
@@ -16,7 +16,7 @@
   -->
 
 <view xmlns="http://jmix.io/schema/flowui/view"
-      title="msg://emailTokenView.title">
+      title="msg://emailConnectionView.title">
     <data>
         <instance id="refreshTokenDc" class="io.jmix.email.entity.RefreshToken">
             <loader id="refreshTokenDl">
